ComboBox added, multiple drivers are allowed
TextBox for location/path was changed to ComboBox to allow much simpler moving between drivers. Double clicking on a file now does not add it to the stack making going back to the same location as you are currently in impossible. Allowing user to use enter key to open directories